requirements.txt
clip==1.0
filelock==3.15.4
fsspec==2024.6.1
ftfy==6.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.0
packaging==24.1
pillow==10.4.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
regex==2024.5.15
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
wcwidth==0.2.13
